refactor: Converter 폴더 controller로 이동
